All URLs treated as equal
Rolled back to primary version, treating each tab or window as equal, no
matter what URL. There was a loophole where a user could open a
chrome:// page and then visit any other page, even if it matches the
forbidden URL patterns. So until there is a fix for this loophole,
extension is dumbed down.